get userwithjobs and jobwithuser

+++ routes/user.py
@@ -1,18 +1,21 @@
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.future import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from db import get_session
-from models import User, UserBase
+from models import User, UserBase, UserwithJobs
+from typing import List, Union
+from sqlalchemy.orm import selectinload
+
 router = APIRouter()
 
-@router.get("/")
+@router.get("/", response_model=List[User])
 async def list_users(session: AsyncSession = Depends(get_session)):
     result = await session.execute(select(User))
     users = result.scalars().all()
     return users
 
-@router.post("/")
+@router.post("/", response_model=User)
 async def create_user(user: UserBase, session: AsyncSession = Depends(get_session)):
     dbuser = User(name=user.name, surname=user.surname, email=user.email )
     session.add(dbuser)
@@ -20,17 +23,20 @@ async def create_user(user: UserBase, session: AsyncSession = Depends(get_sessio
     await session.refresh(dbuser)
     return dbuser
 
-@router.get("/{user_id}")
+@router.get("/{user_id}", response_model=UserwithJobs)
 async def get_user(user_id: int, session: AsyncSession = Depends(get_session)):
-    result = await session.execute(select(User).filter(User.id == user_id))
+    result = await session.execute(select(User).filter(User.id == user_id).options(selectinload(User.jobs)))
     user = result.scalars().first()
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
     return user
 
-@router.put("/{user_id}")
+@router.put("/{user_id}", response_model=User)
 async def update_user(user_id: int, user: UserBase, session: AsyncSession = Depends(get_session)):
     result = await session.execute(select(User).filter(User.id == user_id))
     dbuser = result.scalars().first()
-    print(f"User: {dbuser}")
+    if not dbuser:
+        raise HTTPException(status_code=404, detail="User not found")
     dbuser.name = user.name
     dbuser.surname = user.surname
     dbuser.email = user.email
@@ -41,7 +47,8 @@ async def update_user(user_id: int, user: UserBase, session: AsyncSession = Depe
 async def delete_user(user_id: int, session: AsyncSession = Depends(get_session)):
     result = await session.execute(select(User).filter(User.id == user_id))
     dbuser = result.scalars().first()
-    print(f"User: {dbuser}")
+    if not dbuser:
+        raise HTTPException(status_code=404, detail="User not found")
     await session.delete(dbuser)
     await session.commit()
-    return dbuser
+    return  {"ok": "user deleted"}
